Adam/game state

+++ Implementation/core/src/main/java/com/teamonehundred/pixelboat/entities/AIBoat.java
@@ -29,11 +29,11 @@ public class AIBoat extends Boat {
     /**
      * Construct a AIBoat object at point (x,y) with default size, texture and animation.
      *
-     * @param x int coordinate for the bottom left point of the boat
-     * @param y int coordinate for the bottom left point of the boat
+     * @param x float coordinate for the bottom left point of the boat
+     * @param y float coordinate for the bottom left point of the boat
      * @author James Frost
      */
-    public AIBoat(int x, int y) {
+    public AIBoat(float x, float y) {
         super(x, y);
 
         initialise();
